requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pyodbc==5.1.0
requests==2.31.0
Pillow==10.1.0
numpy==1.26.0
torch==2.3.0
torchvision==0.18.0
open-clip-torch==2.24.0
pandas==2.2.0
scikit-learn==1.5.0
matplotlib==3.9.0
